+++ cli/src/main.rs
@@ -8,9 +8,9 @@ mod error;
 mod gas;
 mod prelude;
 mod scheduler;
+mod secrets;
 mod sui;
 mod tool;
-mod utils;
 mod workflow;
 
 use crate::prelude::*;
@@ -47,6 +47,8 @@ enum Command {
     Gas(gas::GasCommand),
     #[command(subcommand, about = "Manage Nexus crypto")]
     Crypto(crypto::CryptoCommand),
+    #[command(subcommand, about = "Manage local secret storage")]
+    Secrets(secrets::SecretsCommand),
     #[command(about = "Provide shell completions")]
     Completion(completion::CompletionCommand),
 }
@@ -92,6 +94,7 @@ async fn main() {
         Command::Gas(gas) => gas::handle(gas).await,
         Command::Scheduler(scheduler) => scheduler::handle(scheduler).await,
         Command::Crypto(crypto) => crypto::handle(crypto).await,
+        Command::Secrets(secrets) => secrets::handle(secrets).await,
         Command::Completion(completion) => completion::handle(completion),
     };